Separa la implementación de la fachada de la implementación de los métodos originales de la librería
